Docusaurus Pipeline: update scripts and templates to handle CI and local flows

+++ pipelines/Docusaurus/templates/build-docs.sh
@@ -2,6 +2,14 @@
 
 scripts=$(dirname "$(realpath $0)")
 
+# Lazily parse args
+if [[ "$1" == "--skip-build" || "$2" == "--skip-build" ]]; then 
+    skip_build="yes"
+fi
+
+if [[ "$1" == "--skip-templates" || "$2" == "--skip-templates" ]]; then 
+    skip_template="yes"
+fi
 
 # We assume that the docs folder is mounted under CURRENT directory.
 # Container will try to use /docs (easy for local builds)
@@ -23,11 +31,18 @@ if [ -f site-config.json ]; then
         exit 1
     fi
 
-    # Move the templates and place release-notes in src/pages for the build
-    cp -r /templates/* .
+    # if running CI/local job, move the templates and place release-notes in src/pages for the build
+    if [[ -d /templates && "${skip_template}" != "yes" ]]; then
+        cp -r /templates/* .
+    fi
     cp release-notes.md "${release_notes}"
 
     # parse templates
+    # if testing, and you've forgot to copy templates, error out
+    if [ ! -f package.json.template ]; then
+        echo "If you're running locally, copy the templates here first and rerun. Otherwise, something went wrong, talk to CI people"
+        exit 1
+    fi
     eval "$(jq -r 'to_entries[] | "export \(.key)=\"\(.value)\""' site-config.json)"
     sed -e "s/{title}/${title}/g" -e "s/{slug}/${slug}/g" -e "s/{projectName}/${projectName}/g" $scripts/docusaurus.config.js.template > ./docusaurus.config.js
     sed -e "s/{projectName}/${projectName}/g" $scripts/package.json.template > ./package.json
@@ -39,7 +54,7 @@ else
     echo "Building docs with Docusaurus2"
 fi
 
-if [ "$1" != "--skip-build" ]; then 
+if [ "${skip_build}" != "yes" ]; then 
     npm ci
     npm run build
 fi
